fix(abac): drop stale-client ID reclaim
Empty ids on PUT are creates, uniqueness is incoming-only, and IDs mint
on write rather than GET. One migration marker; UUID format stays 400.
