+++ api/api_config_test.go
@@ -28,7 +28,7 @@ type testConfigStore struct {
 	saveErr error
 
 	// serviceIDMigrationDone mirrors the store's migration marker driving the
-	// stale legacy UUID rejection in UpdateConfig.
+	// post-migration UUID format rejection in UpdateConfig.
 	serviceIDMigrationDone bool
 }
 
@@ -59,7 +59,7 @@ func (s *testConfigStore) UpdateConfig(transform func(prev *config.Config) (conf
 	if s.serviceIDMigrationDone {
 		for i := range next.Services {
 			if len(next.Services[i].ID) == 36 {
-				return next, store.ErrStaleLegacyServiceIDs
+				return next, store.ErrLegacyUUIDServiceID
 			}
 		}
 	}
@@ -156,6 +156,31 @@ func TestHandleGetConfig(t *testing.T) {
 				assert.Empty(t, cfg.DefaultBotName)
 				assert.True(t, cfg.MCP.Enabled)
 				assert.True(t, cfg.MCP.EmbeddedServer.Enabled)
+				assert.Empty(t, cfg.MCP.EmbeddedServer.ID, "GET must not mint unpersisted IDs")
+			},
+		},
+		{
+			name: "does not mint IDs for ID-less stored rows",
+			storedConfig: &config.Config{
+				Services: []llm.ServiceConfig{
+					{Name: "OpenAI", Type: "openai"},
+				},
+				MCP: config.MCPConfig{
+					Servers: []config.MCPServerConfig{
+						{Name: "Jira", BaseURL: "https://jira.example.com"},
+					},
+				},
+			},
+			expectedStatus: http.StatusOK,
+			validateBody: func(t *testing.T, body []byte) {
+				var cfg config.Config
+				err := json.Unmarshal(body, &cfg)
+				require.NoError(t, err)
+				require.Len(t, cfg.Services, 1)
+				assert.Empty(t, cfg.Services[0].ID)
+				require.Len(t, cfg.MCP.Servers, 1)
+				assert.Empty(t, cfg.MCP.Servers[0].ID)
+				assert.Empty(t, cfg.MCP.EmbeddedServer.ID)
 			},
 		},
 		{
@@ -464,551 +489,6 @@ func TestHandleSaveConfig(t *testing.T) {
 	}
 }
 
-func TestNormalizeAdminConfigAssignsStableIDs(t *testing.T) {
-	tests := []struct {
-		name     string
-		cfg      config.Config
-		validate func(t *testing.T, result config.Config)
-	}{
-		{
-			name: "empty service and MCP server IDs get fresh valid IDs",
-			cfg: config.Config{
-				Services: []llm.ServiceConfig{
-					{Name: "no-id"},
-				},
-				MCP: config.MCPConfig{
-					Servers: []config.MCPServerConfig{
-						{Name: "srv-no-id", BaseURL: "https://one.example.com"},
-					},
-					PluginServers: []config.PluginServerConfig{
-						{PluginID: "com.example.a", Name: "A", Path: "/mcp"},
-					},
-				},
-			},
-			validate: func(t *testing.T, result config.Config) {
-				assert.True(t, model.IsValidId(result.Services[0].ID))
-				assert.True(t, model.IsValidId(result.MCP.Servers[0].ID))
-				assert.True(t, model.IsValidId(result.MCP.EmbeddedServer.ID))
-				assert.True(t, model.IsValidId(result.MCP.PluginServers[0].ID))
-			},
-		},
-		{
-			name: "non-empty IDs untouched",
-			cfg: config.Config{
-				Services: []llm.ServiceConfig{
-					{ID: "svc-existing", Name: "has-id"},
-				},
-				MCP: config.MCPConfig{
-					Servers: []config.MCPServerConfig{
-						{ID: "mcp-existing", Name: "srv-has-id", BaseURL: "https://one.example.com"},
-					},
-					EmbeddedServer: config.MCPEmbeddedServerConfig{ID: "embedded-existing", Enabled: true},
-					PluginServers: []config.PluginServerConfig{
-						{ID: "plugin-existing", PluginID: "com.example.a", Name: "A", Path: "/mcp"},
-					},
-				},
-			},
-			validate: func(t *testing.T, result config.Config) {
-				assert.Equal(t, "svc-existing", result.Services[0].ID)
-				assert.Equal(t, "mcp-existing", result.MCP.Servers[0].ID)
-				assert.Equal(t, "embedded-existing", result.MCP.EmbeddedServer.ID)
-				assert.Equal(t, "plugin-existing", result.MCP.PluginServers[0].ID)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tt.validate(t, normalizeAdminConfig(tt.cfg))
-		})
-	}
-}
-
-func TestHandleSaveConfigCarriesForwardMCPServerIDs(t *testing.T) {
-	tests := []struct {
-		name       string
-		storedCfg  *config.Config
-		payloadCfg config.Config
-		validate   func(t *testing.T, store *testConfigStore)
-	}{
-		{
-			name: "payload without id keeps stored id (stale bundle)",
-			storedCfg: &config.Config{
-				MCP: config.MCPConfig{
-					Servers: []config.MCPServerConfig{
-						{ID: "stable-id", Name: "Jira", BaseURL: "https://jira.example.com"},
-					},
-					EmbeddedServer: config.MCPEmbeddedServerConfig{ID: "embedded-stable", Enabled: true},
-					PluginServers: []config.PluginServerConfig{
-						{ID: "plugin-stable", PluginID: "com.example.a", Name: "A", Path: "/mcp"},
-					},
-				},
-			},
-			payloadCfg: config.Config{
-				MCP: config.MCPConfig{
-					Servers: []config.MCPServerConfig{
-						{Name: "Jira", BaseURL: "https://jira.example.com"},
-					},
-					EmbeddedServer: config.MCPEmbeddedServerConfig{Enabled: true},
-					PluginServers: []config.PluginServerConfig{
-						{PluginID: "com.example.a", Name: "A", Path: "/mcp/v2"},
-					},
-				},
-			},
-			validate: func(t *testing.T, store *testConfigStore) {
-				require.Len(t, store.cfg.MCP.Servers, 1)
-				assert.Equal(t, "stable-id", store.cfg.MCP.Servers[0].ID)
-				assert.Equal(t, "embedded-stable", store.cfg.MCP.EmbeddedServer.ID)
-				require.Len(t, store.cfg.MCP.PluginServers, 1)
-				assert.Equal(t, "plugin-stable", store.cfg.MCP.PluginServers[0].ID)
-			},
-		},
-		{
-			name: "renamed server without id matched by BaseURL",
-			storedCfg: &config.Config{
-				MCP: config.MCPConfig{
-					Servers: []config.MCPServerConfig{
-						{ID: "stable-id", Name: "Old Name", BaseURL: "https://jira.example.com"},
-					},
-				},
-			},
-			payloadCfg: config.Config{
-				MCP: config.MCPConfig{
-					Servers: []config.MCPServerConfig{
-						{Name: "New Name", BaseURL: "https://jira.example.com"},
-					},
-				},
-			},
-			validate: func(t *testing.T, store *testConfigStore) {
-				require.Len(t, store.cfg.MCP.Servers, 1)
-				assert.Equal(t, "stable-id", store.cfg.MCP.Servers[0].ID)
-			},
-		},
-		{
-			name: "brand-new server without match gets a fresh id from the backstop",
-			storedCfg: &config.Config{
-				MCP: config.MCPConfig{
-					Servers: []config.MCPServerConfig{
-						{ID: "stable-id", Name: "Jira", BaseURL: "https://jira.example.com"},
-					},
-				},
-			},
-			payloadCfg: config.Config{
-				MCP: config.MCPConfig{
-					Servers: []config.MCPServerConfig{
-						{Name: "Jira", BaseURL: "https://jira.example.com"},
-						{Name: "Brand New", BaseURL: "https://new.example.com"},
-					},
-				},
-			},
-			validate: func(t *testing.T, store *testConfigStore) {
-				require.Len(t, store.cfg.MCP.Servers, 2)
-				assert.Equal(t, "stable-id", store.cfg.MCP.Servers[0].ID)
-				assert.True(t, model.IsValidId(store.cfg.MCP.Servers[1].ID))
-				assert.NotEqual(t, "stable-id", store.cfg.MCP.Servers[1].ID)
-			},
-		},
-		{
-			name:      "no stored config still assigns ids",
-			storedCfg: nil,
-			payloadCfg: config.Config{
-				MCP: config.MCPConfig{
-					Servers: []config.MCPServerConfig{
-						{Name: "Jira", BaseURL: "https://jira.example.com"},
-					},
-				},
-			},
-			validate: func(t *testing.T, store *testConfigStore) {
-				require.Len(t, store.cfg.MCP.Servers, 1)
-				assert.True(t, model.IsValidId(store.cfg.MCP.Servers[0].ID))
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			store := &testConfigStore{cfg: tt.storedCfg}
-			router := setupTestRouter(store, &testConfigUpdater{}, &testClusterNotifier{})
-
-			body, err := json.Marshal(tt.payloadCfg)
-			require.NoError(t, err)
-
-			req := httptest.NewRequest(http.MethodPut, "/admin/config", bytes.NewReader(body))
-			req.Header.Set("Content-Type", "application/json")
-			w := httptest.NewRecorder()
-			router.ServeHTTP(w, req)
-
-			require.Equal(t, http.StatusOK, w.Code)
-			tt.validate(t, store)
-		})
-	}
-}
-
-// TestHandleSaveConfigPreservesOmittedPluginServers: full config save never
-// accepts client-provided plugin_servers — omitted and stale non-empty lists
-// both leave persisted plugin rows (and their IDs) untouched.
-func TestHandleSaveConfigPreservesOmittedPluginServers(t *testing.T) {
-	pluginA := "abcdefghijklmnopqrstuvwx0a"
-	pluginB := "abcdefghijklmnopqrstuvwx0b"
-	embeddedID := "abcdefghijklmnopqrstuvwx0e"
-	remoteID := "abcdefghijklmnopqrstuvwx0r"
-
-	storedPlugins := []config.PluginServerConfig{
-		{ID: pluginA, PluginID: "com.example.a", Name: "A", Path: "/mcp", Enabled: true},
-		{ID: pluginB, PluginID: "com.example.b", Name: "B", Path: "/other", Enabled: false},
-	}
-
-	tests := []struct {
-		name    string
-		mcp     map[string]any
-		wantIDs []string
-	}{
-		{
-			name: "omitted plugin_servers preserves prev",
-			mcp: map[string]any{
-				"enabled": true,
-				"servers": []map[string]any{
-					{"name": "Remote", "baseURL": "https://mcp.example.com", "enabled": true},
-				},
-				"embeddedServer": map[string]any{"enabled": true},
-			},
-			wantIDs: []string{pluginA, pluginB},
-		},
-		{
-			name: "stale non-empty plugin_servers ignored",
-			mcp: map[string]any{
-				"enabled": true,
-				"servers": []map[string]any{
-					{"name": "Remote", "baseURL": "https://mcp.example.com", "enabled": true},
-				},
-				"embeddedServer": map[string]any{"enabled": true},
-				"plugin_servers": []map[string]any{
-					{
-						"id": "attackeridattackeridattack", "plugin_id": "com.example.a",
-						"name": "Hacked", "path": "/evil", "enabled": false,
-					},
-				},
-			},
-			wantIDs: []string{pluginA, pluginB},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			store := &testConfigStore{cfg: &config.Config{
-				MCP: config.MCPConfig{
-					Enabled: true,
-					Servers: []config.MCPServerConfig{
-						{ID: remoteID, Name: "Remote", BaseURL: "https://mcp.example.com", Enabled: true},
-					},
-					EmbeddedServer: config.MCPEmbeddedServerConfig{ID: embeddedID, Enabled: true},
-					PluginServers:  append([]config.PluginServerConfig(nil), storedPlugins...),
-				},
-			}}
-			router := setupTestRouter(store, &testConfigUpdater{}, &testClusterNotifier{})
-
-			body, err := json.Marshal(map[string]any{"mcp": tt.mcp})
-			require.NoError(t, err)
-
-			req := httptest.NewRequest(http.MethodPut, "/admin/config", bytes.NewReader(body))
-			req.Header.Set("Content-Type", "application/json")
-			w := httptest.NewRecorder()
-			router.ServeHTTP(w, req)
-			require.Equal(t, http.StatusOK, w.Code)
-
-			require.Len(t, store.cfg.MCP.PluginServers, len(tt.wantIDs))
-			for i, id := range tt.wantIDs {
-				assert.Equal(t, id, store.cfg.MCP.PluginServers[i].ID)
-				assert.Equal(t, storedPlugins[i].PluginID, store.cfg.MCP.PluginServers[i].PluginID)
-				assert.Equal(t, storedPlugins[i].Name, store.cfg.MCP.PluginServers[i].Name)
-				assert.Equal(t, storedPlugins[i].Path, store.cfg.MCP.PluginServers[i].Path)
-				assert.Equal(t, storedPlugins[i].Enabled, store.cfg.MCP.PluginServers[i].Enabled)
-			}
-			assert.Equal(t, remoteID, store.cfg.MCP.Servers[0].ID)
-			assert.Equal(t, embeddedID, store.cfg.MCP.EmbeddedServer.ID)
-		})
-	}
-}
-
-func TestHandleSaveConfigRejectsCrossKindMCPIDConflict(t *testing.T) {
-	sharedID := "abcdefghijklmnopqrstuvwxzz"
-	store := &testConfigStore{cfg: &config.Config{
-		MCP: config.MCPConfig{
-			EmbeddedServer: config.MCPEmbeddedServerConfig{ID: sharedID, Enabled: true},
-		},
-	}}
-	router := setupTestRouter(store, &testConfigUpdater{}, &testClusterNotifier{})
-
-	payload := config.Config{
-		MCP: config.MCPConfig{
-			Servers: []config.MCPServerConfig{
-				{ID: sharedID, Name: "Remote", BaseURL: "https://mcp.example.com", Enabled: true},
-			},
-			EmbeddedServer: config.MCPEmbeddedServerConfig{ID: sharedID, Enabled: true},
-		},
-	}
-	body, err := json.Marshal(payload)
-	require.NoError(t, err)
-
-	req := httptest.NewRequest(http.MethodPut, "/admin/config", bytes.NewReader(body))
-	req.Header.Set("Content-Type", "application/json")
-	w := httptest.NewRecorder()
-	router.ServeHTTP(w, req)
-	require.Equal(t, http.StatusConflict, w.Code)
-}
-
-// TestHandleSaveConfigCarriesForwardServiceIDs mirrors the MCP server ID
-// reconciliation for LLM services: service IDs are ABAC policy identities, so
-// an ID-less payload from a stale client must reclaim the stored ID instead
-// of rotating it (which would silently detach the service's policy), and
-// unresolvable identities must fail with 409.
-func TestHandleSaveConfigCarriesForwardServiceIDs(t *testing.T) {
-	stored := func() *config.Config {
-		return &config.Config{
-			Services: []llm.ServiceConfig{
-				{ID: "stable-svc-id", Name: "OpenAI", Type: "openai"},
-			},
-		}
-	}
-
-	tests := []struct {
-		name            string
-		storedCfg       *config.Config
-		payloadServices []llm.ServiceConfig
-		expectedStatus  int
-		validate        func(t *testing.T, store *testConfigStore)
-	}{
-		{
-			name:      "payload without id keeps stored id (stale client)",
-			storedCfg: stored(),
-			payloadServices: []llm.ServiceConfig{
-				{Name: "OpenAI", Type: "openai"},
-			},
-			expectedStatus: http.StatusOK,
-			validate: func(t *testing.T, store *testConfigStore) {
-				require.Len(t, store.cfg.Services, 1)
-				assert.Equal(t, "stable-svc-id", store.cfg.Services[0].ID)
-			},
-		},
-		{
-			name:      "repeated ID-less saves never rotate the id",
-			storedCfg: stored(),
-			payloadServices: []llm.ServiceConfig{
-				{Name: "OpenAI", Type: "anthropic", APIURL: "https://edited.example.com"},
-			},
-			expectedStatus: http.StatusOK,
-			validate: func(t *testing.T, store *testConfigStore) {
-				require.Len(t, store.cfg.Services, 1)
-				assert.Equal(t, "stable-svc-id", store.cfg.Services[0].ID)
-			},
-		},
-		{
-			name:      "brand-new service without match gets a fresh id from the backstop",
-			storedCfg: stored(),
-			payloadServices: []llm.ServiceConfig{
-				{ID: "stable-svc-id", Name: "OpenAI", Type: "openai"},
-				{Name: "Brand New", Type: "anthropic"},
-			},
-			expectedStatus: http.StatusOK,
-			validate: func(t *testing.T, store *testConfigStore) {
-				require.Len(t, store.cfg.Services, 2)
-				assert.Equal(t, "stable-svc-id", store.cfg.Services[0].ID)
-				assert.True(t, model.IsValidId(store.cfg.Services[1].ID))
-			},
-		},
-		{
-			name:      "fabricated incoming ID returns 409",
-			storedCfg: stored(),
-			payloadServices: []llm.ServiceConfig{
-				{ID: "invented-by-client", Name: "OpenAI", Type: "openai"},
-			},
-			expectedStatus: http.StatusConflict,
-			validate: func(t *testing.T, store *testConfigStore) {
-				assert.Equal(t, "stable-svc-id", store.cfg.Services[0].ID, "stored config must be untouched on rejection")
-			},
-		},
-		{
-			name:      "no stored config still assigns ids",
-			storedCfg: nil,
-			payloadServices: []llm.ServiceConfig{
-				{Name: "OpenAI", Type: "openai"},
-			},
-			expectedStatus: http.StatusOK,
-			validate: func(t *testing.T, store *testConfigStore) {
-				require.Len(t, store.cfg.Services, 1)
-				assert.True(t, model.IsValidId(store.cfg.Services[0].ID))
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			store := &testConfigStore{cfg: tt.storedCfg}
-			router := setupTestRouter(store, &testConfigUpdater{}, &testClusterNotifier{})
-
-			body, err := json.Marshal(config.Config{Services: tt.payloadServices})
-			require.NoError(t, err)
-
-			req := httptest.NewRequest(http.MethodPut, "/admin/config", bytes.NewReader(body))
-			req.Header.Set("Content-Type", "application/json")
-			w := httptest.NewRecorder()
-			router.ServeHTTP(w, req)
-
-			require.Equal(t, tt.expectedStatus, w.Code)
-			if tt.validate != nil {
-				tt.validate(t, store)
-			}
-		})
-	}
-}
-
-// TestHandleSaveConfigRejectsMCPServerIDConflicts covers stale or corrupt
-// admin payloads whose MCP server identities cannot be reconciled with the
-// stored config: the save must fail with 409 (never silently re-mint IDs,
-// which would detach existing policies), while legitimate add/edit flows
-// still succeed.
-func TestHandleSaveConfigRejectsMCPServerIDConflicts(t *testing.T) {
-	storedServer := config.MCPServerConfig{ID: "stable-id", Name: "Jira", BaseURL: "https://jira.example.com"}
-	storedOther := config.MCPServerConfig{ID: "other-id", Name: "Linear", BaseURL: "https://linear.example.com"}
-
-	tests := []struct {
-		name           string
-		payloadServers []config.MCPServerConfig
-		expectedStatus int
-		validate       func(t *testing.T, store *testConfigStore)
-	}{
-		{
-			name: "fabricated incoming ID returns 409",
-			payloadServers: []config.MCPServerConfig{
-				{ID: "invented-by-client", Name: "Jira", BaseURL: "https://jira.example.com"},
-			},
-			expectedStatus: http.StatusConflict,
-		},
-		{
-			name: "duplicate incoming IDs return 409",
-			payloadServers: []config.MCPServerConfig{
-				{ID: "stable-id", Name: "Jira", BaseURL: "https://jira.example.com"},
-				{ID: "stable-id", Name: "Copy", BaseURL: "https://copy.example.com"},
-			},
-			expectedStatus: http.StatusConflict,
-		},
-		{
-			name: "ambiguous ID-less identity match returns 409",
-			payloadServers: []config.MCPServerConfig{
-				// Name matches Linear, BaseURL matches Jira: never guess.
-				{Name: "Linear", BaseURL: "https://jira.example.com"},
-			},
-			expectedStatus: http.StatusConflict,
-		},
-		{
-			name: "add flow: new ID-less server minted server-side",
-			payloadServers: []config.MCPServerConfig{
-				{ID: "stable-id", Name: "Jira", BaseURL: "https://jira.example.com"},
-				{ID: "other-id", Name: "Linear", BaseURL: "https://linear.example.com"},
-				{Name: "Brand New", BaseURL: "https://new.example.com"},
-			},
-			expectedStatus: http.StatusOK,
-			validate: func(t *testing.T, store *testConfigStore) {
-				require.Len(t, store.cfg.MCP.Servers, 3)
-				assert.Equal(t, "stable-id", store.cfg.MCP.Servers[0].ID)
-				assert.Equal(t, "other-id", store.cfg.MCP.Servers[1].ID)
-				assert.True(t, model.IsValidId(store.cfg.MCP.Servers[2].ID), "new server gets a server-minted ID")
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			stored := &config.Config{
-				MCP: config.MCPConfig{Servers: []config.MCPServerConfig{storedServer, storedOther}},
-			}
-			store := &testConfigStore{cfg: stored}
-			updater := &testConfigUpdater{}
-			notifier := &testClusterNotifier{}
-			router := setupTestRouter(store, updater, notifier)
-
-			body, err := json.Marshal(config.Config{
-				MCP: config.MCPConfig{Servers: tt.payloadServers},
-			})
-			require.NoError(t, err)
-
-			req := httptest.NewRequest(http.MethodPut, "/admin/config", bytes.NewReader(body))
-			req.Header.Set("Content-Type", "application/json")
-			w := httptest.NewRecorder()
-			router.ServeHTTP(w, req)
-
-			require.Equal(t, tt.expectedStatus, w.Code)
-			if tt.expectedStatus == http.StatusConflict {
-				assert.Equal(t, "stable-id", store.cfg.MCP.Servers[0].ID, "stored config must be untouched on rejection")
-				assert.Equal(t, 0, updater.callCount, "in-memory config must not be updated on rejection")
-				assert.Equal(t, 0, notifier.callCount, "cluster must not be notified on rejection")
-			}
-			if tt.validate != nil {
-				tt.validate(t, store)
-			}
-		})
-	}
-}
-
-// TestHandleSaveConfigFirstWriteIncomingMCPIDs covers the first-write path
-// (no stored config yet): reconciliation must still run against an empty
-// previous list, so duplicate incoming MCP server IDs are rejected, while
-// caller-chosen IDs for genuinely new servers (API automation seeding a
-// fresh install) are kept.
-func TestHandleSaveConfigFirstWriteIncomingMCPIDs(t *testing.T) {
-	tests := []struct {
-		name           string
-		payloadServers []config.MCPServerConfig
-		expectedStatus int
-	}{
-		{
-			name: "caller-chosen ID on first write is kept",
-			payloadServers: []config.MCPServerConfig{
-				{ID: "seeded-by-client", Name: "Jira", BaseURL: "https://jira.example.com"},
-			},
-			expectedStatus: http.StatusOK,
-		},
-		{
-			name: "duplicate IDs on first write",
-			payloadServers: []config.MCPServerConfig{
-				{ID: "seeded-by-client", Name: "Jira", BaseURL: "https://jira.example.com"},
-				{ID: "seeded-by-client", Name: "Copy", BaseURL: "https://copy.example.com"},
-			},
-			expectedStatus: http.StatusConflict,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			store := &testConfigStore{cfg: nil}
-			updater := &testConfigUpdater{}
-			notifier := &testClusterNotifier{}
-			router := setupTestRouter(store, updater, notifier)
-
-			body, err := json.Marshal(config.Config{
-				MCP: config.MCPConfig{Servers: tt.payloadServers},
-			})
-			require.NoError(t, err)
-
-			req := httptest.NewRequest(http.MethodPut, "/admin/config", bytes.NewReader(body))
-			req.Header.Set("Content-Type", "application/json")
-			w := httptest.NewRecorder()
-			router.ServeHTTP(w, req)
-
-			require.Equal(t, tt.expectedStatus, w.Code)
-			if tt.expectedStatus == http.StatusConflict {
-				assert.Nil(t, store.cfg, "nothing must be persisted on rejection")
-				assert.Equal(t, 0, updater.callCount)
-				assert.Equal(t, 0, notifier.callCount)
-			} else {
-				require.NotNil(t, store.cfg)
-				require.Len(t, store.cfg.MCP.Servers, len(tt.payloadServers))
-				assert.Equal(t, tt.payloadServers[0].ID, store.cfg.MCP.Servers[0].ID)
-			}
-		})
-	}
-}
-
 // TestHandleSaveConfigReturnsNormalizedConfig verifies the PUT response body
 // carries the normalized saved config, so the webapp can adopt server-minted
 // service and MCP server IDs immediately instead of waiting for a reload.
@@ -1045,56 +525,6 @@ func TestHandleSaveConfigReturnsNormalizedConfig(t *testing.T) {
 	assert.True(t, resp.Services[0].UseResponsesAPI, "response must reflect normalization")
 }
 
-// TestHandleSaveConfigRejectsStaleLegacyServiceIDs covers the interleaving
-// where a pre-upgrade webapp bundle loaded the config before the ID migration
-// ran and then saves UUID service IDs back: the save must fail with 409 and
-// leave the migrated config untouched, while a fresh payload is accepted.
-func TestHandleSaveConfigRejectsStaleLegacyServiceIDs(t *testing.T) {
-	migrated := &config.Config{
-		Services: []llm.ServiceConfig{
-			{ID: "migrated26charidmigrated26", Name: "OpenAI", Type: "openai"},
-		},
-	}
-	store := &testConfigStore{cfg: migrated, serviceIDMigrationDone: true}
-	updater := &testConfigUpdater{}
-	notifier := &testClusterNotifier{}
-	router := setupTestRouter(store, updater, notifier)
-
-	put := func(cfg config.Config) *httptest.ResponseRecorder {
-		body, err := json.Marshal(cfg)
-		require.NoError(t, err)
-		req := httptest.NewRequest(http.MethodPut, "/admin/config", bytes.NewReader(body))
-		req.Header.Set("Content-Type", "application/json")
-		w := httptest.NewRecorder()
-		router.ServeHTTP(w, req)
-		return w
-	}
-
-	// Stale payload echoing a pre-migration UUID service ID.
-	stale := config.Config{
-		Services: []llm.ServiceConfig{
-			{ID: "550e8400-e29b-41d4-a716-446655440000", Name: "OpenAI", Type: "openai"},
-		},
-	}
-	w := put(stale)
-	assert.Equal(t, http.StatusConflict, w.Code)
-	assert.Equal(t, "migrated26charidmigrated26", store.cfg.Services[0].ID, "migrated ID must survive the stale save")
-	assert.Equal(t, 0, updater.callCount, "in-memory config must not be updated on rejection")
-	assert.Equal(t, 0, notifier.callCount, "cluster must not be notified on rejection")
-
-	// Fresh payload with the migrated ID is accepted.
-	fresh := config.Config{
-		Services: []llm.ServiceConfig{
-			{ID: "migrated26charidmigrated26", Name: "OpenAI Renamed", Type: "openai"},
-		},
-	}
-	w = put(fresh)
-	assert.Equal(t, http.StatusOK, w.Code)
-	assert.Equal(t, "OpenAI Renamed", store.cfg.Services[0].Name)
-	assert.Equal(t, 1, updater.callCount)
-	assert.Equal(t, 1, notifier.callCount)
-}
-
 func TestSaveAndGetConfigRoundTrip(t *testing.T) {
 	store := &testConfigStore{}
 	updater := &testConfigUpdater{}
@@ -1113,7 +543,7 @@ func TestSaveAndGetConfigRoundTrip(t *testing.T) {
 	assert.Empty(t, emptyCfg.Services)
 
 	// Step 2: PUT a config. The new service arrives ID-less (the backend
-	// mints the stable ID); an unknown incoming ID would be rejected.
+	// mints the stable ID).
 	saveCfg := config.Config{
 		DefaultBotName: "ai",
 		Services: []llm.ServiceConfig{
