+++ config/mcp_config_test.go
@@ -244,268 +244,6 @@ func TestMCPConfigServerIDMaps(t *testing.T) {
 	}
 }
 
-func TestReconcileMCPServerIDs(t *testing.T) {
-	tests := []struct {
-		name      string
-		next      []MCPServerConfig
-		prev      []MCPServerConfig
-		expectIDs []string
-		expectErr bool
-	}{
-		{
-			name: "round-trip payload keeps existing ID",
-			next: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			expectIDs: []string{"prev-id"},
-		},
-		{
-			name: "match by name when URL edited",
-			next: []MCPServerConfig{
-				{Name: "srv", BaseURL: "https://new-url.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://old-url.example.com"},
-			},
-			expectIDs: []string{"prev-id"},
-		},
-		{
-			name: "match by BaseURL when renamed",
-			next: []MCPServerConfig{
-				{Name: "renamed", BaseURL: "https://one.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			expectIDs: []string{"prev-id"},
-		},
-		{
-			name: "each prev entry consumed at most once",
-			next: []MCPServerConfig{
-				{Name: "srv", BaseURL: "https://one.example.com"},
-				{Name: "other", BaseURL: "https://one.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			expectIDs: []string{"prev-id", ""},
-		},
-		{
-			name: "add flow: genuinely new entry stays ID-less for the caller to mint",
-			next: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://old.example.com"},
-				{Name: "brand-new", BaseURL: "https://new.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://old.example.com"},
-			},
-			expectIDs: []string{"prev-id", ""},
-		},
-		{
-			name: "delete flow: fewer entries than prev is not an error",
-			next: []MCPServerConfig{
-				{ID: "id-one", Name: "keep", BaseURL: "https://one.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "id-one", Name: "keep", BaseURL: "https://one.example.com"},
-				{ID: "id-two", Name: "delete-me", BaseURL: "https://two.example.com"},
-			},
-			expectIDs: []string{"id-one"},
-		},
-		{
-			name: "prev entries without IDs cannot be claimed",
-			next: []MCPServerConfig{
-				{Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			expectIDs: []string{""},
-		},
-		{
-			name: "duplicate names reordered resolve by exact (Name, BaseURL)",
-			next: []MCPServerConfig{
-				{Name: "dup", BaseURL: "https://two.example.com"},
-				{Name: "dup", BaseURL: "https://one.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "id-one", Name: "dup", BaseURL: "https://one.example.com"},
-				{ID: "id-two", Name: "dup", BaseURL: "https://two.example.com"},
-			},
-			expectIDs: []string{"id-two", "id-one"},
-		},
-		{
-			name: "explicit-ID claim and exact match competing across phases rejected",
-			next: []MCPServerConfig{
-				// The ID-less entry exactly matches the stored server the
-				// ID-bearing entry still holds: two entries strongly claiming
-				// one identity is a corrupt payload, never a guess.
-				{Name: "srv", BaseURL: "https://one.example.com"},
-				{ID: "prev-id", Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			expectErr: true,
-		},
-		{
-			name: "weak match earlier in the payload cannot steal a later exact match",
-			next: []MCPServerConfig{
-				// Payload order: the unique-Name weak match comes first, but
-				// the second entry is an exact (Name, BaseURL) round-trip of
-				// the stored server. Exact claims resolve globally before any
-				// weak claim, so the first entry stays new.
-				{Name: "srv", BaseURL: "https://moved.example.com"},
-				{Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			expectIDs: []string{"", "prev-id"},
-		},
-		{
-			name: "two identical ID-less entries competing for one stored server rejected",
-			next: []MCPServerConfig{
-				{Name: "srv", BaseURL: "https://one.example.com"},
-				{Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			expectErr: true,
-		},
-		{
-			name: "two weak claims resolving to the same stored server rejected",
-			next: []MCPServerConfig{
-				// First entry matches by unique Name, second by unique
-				// BaseURL — both resolve to the single stored server, and
-				// which one deserves the identity cannot be decided.
-				{Name: "srv", BaseURL: "https://a.example.com"},
-				{Name: "other", BaseURL: "https://one.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			expectErr: true,
-		},
-		{
-			name: "first write: ID-less entries stay ID-less for the caller to mint",
-			next: []MCPServerConfig{
-				{Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			prev:      nil,
-			expectIDs: []string{""},
-		},
-		{
-			name: "duplicate incoming IDs rejected",
-			next: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://one.example.com"},
-				{ID: "prev-id", Name: "copy", BaseURL: "https://two.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			expectErr: true,
-		},
-		{
-			name: "unknown ID colliding with a stored server identity rejected",
-			next: []MCPServerConfig{
-				{ID: "never-issued-id", Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			expectErr: true,
-		},
-		{
-			name: "unknown ID with unique name and URL is kept (API automation)",
-			next: []MCPServerConfig{
-				{ID: "seeded-by-automation", Name: "new-srv", BaseURL: "https://new.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "prev-id", Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			expectIDs: []string{"seeded-by-automation"},
-		},
-		{
-			name: "caller-chosen IDs on first write are kept",
-			next: []MCPServerConfig{
-				{ID: "seeded-by-automation", Name: "srv", BaseURL: "https://one.example.com"},
-			},
-			prev:      nil,
-			expectIDs: []string{"seeded-by-automation"},
-		},
-		{
-			name: "rename plus URL-change crossing two prev servers rejected",
-			next: []MCPServerConfig{
-				// Name matches prev B, BaseURL matches prev A: never guess.
-				{Name: "beta", BaseURL: "https://alpha.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "id-alpha", Name: "alpha", BaseURL: "https://alpha.example.com"},
-				{ID: "id-beta", Name: "beta", BaseURL: "https://beta.example.com"},
-			},
-			expectErr: true,
-		},
-		{
-			name: "multiple name matches with no exact match rejected",
-			next: []MCPServerConfig{
-				{Name: "dup", BaseURL: "https://three.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "id-one", Name: "dup", BaseURL: "https://one.example.com"},
-				{ID: "id-two", Name: "dup", BaseURL: "https://two.example.com"},
-			},
-			expectErr: true,
-		},
-		{
-			name: "duplicate non-empty stored IDs rejected",
-			next: []MCPServerConfig{
-				// One entry claims the ID explicitly, the other by exact
-				// (Name, BaseURL) match: with two stored rows sharing the ID,
-				// both claims could be satisfied by different rows, leaving
-				// two servers guarded by one policy identity.
-				{ID: "shared-id", Name: "srv", BaseURL: "https://one.example.com"},
-				{Name: "copy", BaseURL: "https://two.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "shared-id", Name: "srv", BaseURL: "https://one.example.com"},
-				{ID: "shared-id", Name: "copy", BaseURL: "https://two.example.com"},
-			},
-			expectErr: true,
-		},
-		{
-			name: "multiple exact (Name, BaseURL) duplicates in prev rejected",
-			next: []MCPServerConfig{
-				{Name: "dup", BaseURL: "https://one.example.com"},
-			},
-			prev: []MCPServerConfig{
-				{ID: "id-one", Name: "dup", BaseURL: "https://one.example.com"},
-				{ID: "id-two", Name: "dup", BaseURL: "https://one.example.com"},
-			},
-			expectErr: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result, err := ReconcileMCPServerIDs(tt.next, tt.prev)
-			if tt.expectErr {
-				require.ErrorIs(t, err, ErrMCPServerIDConflict)
-				return
-			}
-			require.NoError(t, err)
-			require.Len(t, result, len(tt.expectIDs))
-			for i, id := range tt.expectIDs {
-				require.Equal(t, id, result[i].ID, "server %d", i)
-			}
-		})
-	}
-}
-
 func TestReconcileEmbeddedMCPServerID(t *testing.T) {
 	tests := []struct {
 		name      string
@@ -575,9 +313,9 @@ func TestReconcileMCPConfigIDs(t *testing.T) {
 		validate  func(t *testing.T, got MCPConfig)
 	}{
 		{
-			name: "happy path carries remote/embedded IDs and prev plugin rows",
+			name: "happy path keeps explicit remote ID, copies embedded/plugin from prev",
 			next: MCPConfig{
-				Servers:        []MCPServerConfig{{Name: "Jira", BaseURL: "https://jira.example.com"}},
+				Servers:        []MCPServerConfig{{ID: "remote-id", Name: "Jira", BaseURL: "https://jira.example.com"}},
 				EmbeddedServer: MCPEmbeddedServerConfig{Enabled: true},
 				PluginServers:  []PluginServerConfig{{PluginID: "com.example.stale", Name: "Stale", Path: "/stale"}},
 			},
@@ -673,6 +411,53 @@ func TestReconcileMCPConfigIDs(t *testing.T) {
 				PluginServers: []PluginServerConfig{{ID: sharedID, PluginID: "com.example.a", Name: "A", Path: "/mcp"}},
 			},
 			expectErr: true,
+		},
+		{
+			name: "duplicate incoming remote IDs rejected",
+			next: MCPConfig{
+				Servers: []MCPServerConfig{
+					{ID: sharedID, Name: "srv", BaseURL: "https://one.example.com"},
+					{ID: sharedID, Name: "copy", BaseURL: "https://two.example.com"},
+				},
+			},
+			expectErr: true,
+		},
+		{
+			name: "unique next succeeds even when prev has duplicate stored remote IDs",
+			next: MCPConfig{
+				Servers: []MCPServerConfig{
+					{ID: "unique-id", Name: "srv", BaseURL: "https://one.example.com"},
+				},
+			},
+			prev: MCPConfig{
+				Servers: []MCPServerConfig{
+					{ID: sharedID, Name: "srv", BaseURL: "https://one.example.com"},
+					{ID: sharedID, Name: "copy", BaseURL: "https://two.example.com"},
+				},
+			},
+			validate: func(t *testing.T, got MCPConfig) {
+				require.Len(t, got.Servers, 1)
+				require.Equal(t, "unique-id", got.Servers[0].ID)
+			},
+		},
+		{
+			name: "ID-less remotes stay empty for the caller to mint",
+			next: MCPConfig{
+				Servers: []MCPServerConfig{
+					{Name: "srv", BaseURL: "https://one.example.com"},
+					{Name: "brand-new", BaseURL: "https://new.example.com"},
+				},
+			},
+			prev: MCPConfig{
+				Servers: []MCPServerConfig{
+					{ID: "prev-id", Name: "srv", BaseURL: "https://one.example.com"},
+				},
+			},
+			validate: func(t *testing.T, got MCPConfig) {
+				require.Len(t, got.Servers, 2)
+				require.Empty(t, got.Servers[0].ID)
+				require.Empty(t, got.Servers[1].ID)
+			},
 		},
 	}
